+++ AutoHotkeyx.vcxproj
@@ -217,7 +217,9 @@
     <ClCompile Include="source\os_version.cpp" />
     <ClCompile Include="source\script.cpp" />
     <ClCompile Include="source\script2.cpp" />
-    <ClCompile Include="source\ScriptModules.cpp" />
+    <ClCompile Include="source\ScriptModules\ScriptModules.cpp">
+      <PrecompiledHeaderFile Condition="'$(Configuration)|$(Platform)'=='Release|x64'">..\stdafx.h</PrecompiledHeaderFile>
+    </ClCompile>
     <ClCompile Include="source\script_autoit.cpp" />
     <ClCompile Include="source\script_com.cpp" />
     <ClCompile Include="source\script_expression.cpp" />
@@ -258,9 +260,10 @@
     <ClInclude Include="source\qmath.h" />
     <ClInclude Include="source\resources\resource.h" />
     <ClInclude Include="source\script.h" />
-    <ClInclude Include="source\ScriptModules.h" />
-    <ClInclude Include="source\ScriptModulesAddNames.h" />
-    <ClInclude Include="source\ScriptModulesDefines.h" />
+    <ClInclude Include="source\ScriptModules\ScriptModules.h" />
+    <ClInclude Include="source\ScriptModules\ScriptModulesAddNames.h" />
+    <ClInclude Include="source\ScriptModules\ScriptModulesDefines.h" />
+    <ClInclude Include="source\ScriptModules\ModuleListMethods.h" />
     <ClInclude Include="source\script_com.h" />
     <ClInclude Include="source\script_func_impl.h" />
     <ClInclude Include="source\script_object.h" />
